rename pipelines directories to pipeline_stages
