+++ examples/ramen_rec/app/__init__.py
@@ -1,4 +1,4 @@
 from .utils import *
 from .models import *
-from .pipelines import *
+from .pipeline_stages import *
 from .services import *
