+++ frex/__init__.py
@@ -1,3 +1,3 @@
 from .models import *
-from .pipelines import *
+from .pipeline_stages import *
 from .stores import *
