+++ frex/pipeline_stages/pipeline_executor.py
@@ -1,5 +1,5 @@
 from typing import Tuple, Generator
-from frex.pipelines import _Pipeline
+from frex.pipeline_stages import _Pipeline
 from frex.models import Context, Candidate
